requirements.txt
## Python ==3.12
google-oauth2-tool==0.0.3
google-auth-oauthlib==1.2.1
pymongo==4.11.1
python-dotenv==1.0.1
youtube-transcript-api==0.6.3